requirements.txt
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
requests>=2.31.0
RPi.GPIO>=0.7.1
adafruit-circuitpython-ads1x15>=2.2.23
smbus2>=0.4.0
